multiple: add integrity command
The new integrity command is used to calculate dm-verity data for
the root partition of an image. The dm-verity data come from a separate
fixed size partition that is located before the rootfs partition. This is
auto-discovered by snap-bootstrap using its GPT partition label. The
dm-verity root hash is placed in a manifest.json file which is placed
in the ESP.