refactor connectNbd() to allow manual disconnection
connectNbd() is setting up a cleanup function to disconnect the
device stored in the context state before the program closes. The
new integrity module will need to manually disconnect the device to
perform partitioning operations on the raw image before reconnecting
it again. This creates situations where the connection is manually
closed and then a new one is created without a way to cleanup the
handler that was already setup. To avoid more complicated cleanup logic,
the connection context variable was added as a new variable in the global
context so that it can be updated after a potential reconnection and for
the cleanup handler to be able to refer to that instead.