fix nbdSuite.TestConnectFail
The test is failing because because some called to the mocked udevadm
command are killed before they have time to write to the log file. The
corrupted/imcomplete log file is then parsed by the MockCommand.Call
method that returns an incorrect number of calls even if the command was
in fact called the right number of time.
By adding a synchronization mechanism, we ensure that the call the the
mocked udevadm command has been properly registered before returning
from `qemu-nbd -c ...` (which is also a MockCommand). This makes
tryConnectToDevice kill the udevadm command only after it has written
its logs to the log file.